Fixed merge issues.
