+++ .idea/compiler.xml
@@ -19,7 +19,7 @@
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
-      <module name="MxM" target="1.7" />
+      <module name="MxM" target="1.6" />
     </bytecodeTargetLevel>
   </component>
 </project>
