+++ pom.xml
@@ -22,6 +22,11 @@
             <artifactId>deeplearning4j-core</artifactId>
             <version>0.5.0</version>
         </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>RELEASE</version>
+        </dependency>
     </dependencies>
     <modelVersion>4.0.0</modelVersion>
     <packaging>pom</packaging>
